pinterest-pin-generator/requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# Image processing
Pillow>=10.3.0
pillow-heif>=0.16.0

# Anthropic Claude API (using instead of Gemini)
anthropic>=0.18.0

# Web scraping
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.17.2

# Dropbox API
dropbox==11.36.2

# AWS S3 (for your existing setup)
boto3==1.34.34

# Utilities
colorlog==6.8.2
